Refacto: Controller and View
