+++ src/controller/cli/game_controller.c
@@ -3,11 +3,24 @@
 #include "model/timer.h"
 #include "model/score.h"
 #include "view/cli/game_view.h"
+#include "controller/cli/user_input_handler.h"
+#include "controller/cli/game_win_handler.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdbool.h>
 
-static void gameLoop(Taquin *taquin, Timer *gameTimer, Leaderboard *leaderboard, const char *scoreFilePath);
+static void gameLoop(Taquin *taquin, Timer *gameTimer, Leaderboard *leaderboard, const char *scoreFilePath) {
+    bool gameRunning = true, gameWon = false;
+    while (gameRunning && !checkVictory(taquin)) {
+        displayGameState(taquin, gameTimer);
+        gameRunning = handleUserInput(taquin, &gameWon);
+    }
+
+    if (gameWon) {
+        handleGameWin(gameTimer, leaderboard, scoreFilePath);
+    }
+}
+
 
 void runGame(const char *scoreFilePath) {
     Leaderboard leaderboard = {0};
@@ -25,17 +38,10 @@ void runGame(const char *scoreFilePath) {
 
     gameLoop(taquin, &gameTimer, &leaderboard, scoreFilePath);
 
+    stopTimer(&gameTimer);
+
+    displayLeaderboard(&leaderboard);
+
     freeScores(&leaderboard);
     freeTaquin(taquin);
-}
-
-static void gameLoop(Taquin *taquin, Timer *gameTimer, Leaderboard *leaderboard, const char *scoreFilePath) {
-    bool gameRunning = true, gameWon = false;
-    while (gameRunning) {
-        displayGameState(taquin, gameTimer);
-        gameRunning = handleUserInput(taquin, &gameWon);
-        if (gameWon) {
-            handleGameWin(gameTimer, leaderboard, scoreFilePath);
-        }
-    }
 }
