+++ src/controller/cli/user_input_handler.h
@@ -1,0 +1,9 @@
+#ifndef USER_INPUT_HANDLER_H
+#define USER_INPUT_HANDLER_H
+
+#include "model/taquin.h"
+#include <stdbool.h>
+
+bool handleUserInput(Taquin *taquin, bool *gameWon);
+
+#endif // USER_INPUT_HANDLER_H
